Add IO base class
